# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ServicePro API
  description: |
    # ServicePro API Documentation

    ServicePro is a comprehensive service management platform for field service businesses.
    This API provides endpoints for managing customers, jobs, invoices, quotes, and user authentication.

    ## Authentication

    Most endpoints require authentication using JWT Bearer tokens. Include the token in the
    `Authorization` header:

    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting

    API requests are rate limited to protect the service. Rate limit information is included
    in response headers:

    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

    ## Pagination

    List endpoints support pagination with the following query parameters:

    - `page`: Page number (default: 1)
    - `page_size`: Results per page (default: 20, max: 100)

    Paginated responses include:

    - `total`: Total number of records
    - `page`: Current page number
    - `page_size`: Results per page
    - `total_pages`: Total number of pages

    ## Error Handling

    The API uses standard HTTP status codes and returns errors in a consistent format:

    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Validation failed",
        "details": [
...
        ]
      }
    }
    ```

...
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication token obtained from the login endpoint.
        Include in the Authorization header as: `Bearer <token>`

  # ============================================================================
  # Common Parameters
  # ============================================================================
...
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed per time window
      schema:
        type: integer
      example: 100
    X-RateLimit-Remaining:
      description: Number of requests remaining in current time window
      schema:
        type: integer
      example: 95
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
      example: 1706025600

  # ============================================================================
  # Reusable Schemas
  # ============================================================================
  schemas:
...
    UUID:
      type: string
      format: uuid
      description: Universally Unique Identifier
      example: 550e8400-e29b-41d4-a716-446655440000
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 timestamp
      example: '2024-01-22T10:30:00Z'
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: 'Validation failed'
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
          required:
            - code
            - message
      required:
        - error
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Field that caused the error
          example: 'email'
        message:
          type: string
          description: Error message for this field
          example: 'Invalid email format'
        code:
          type: string
          description: Error code for this field
          example: 'INVALID_FORMAT'
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  example: VALIDATION_ERROR
                message:
                  example: 'Request validation failed'
...
    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  example: RATE_LIMIT_EXCEEDED
                message:
                  example: 'Too many requests. Please try again later.'
                details:
                  type: array
                  items:
                    type: object
                    properties:
                      retry_after:
                        type: integer
                        description: Seconds until rate limit resets
                        example: 60

    # --------------------------------------------------------------------------
    # Authentication Schemas
    # --------------------------------------------------------------------------
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
          example: 'user@example.com'
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: User's password
          example: 'SecureP@ss123'
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
          example: 'dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...'
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        token_type:
          type: string
          enum: [Bearer]
          example: 'Bearer'
        user:
          $ref: '#/components/schemas/UserProfile'
      required:
        - token
        - refresh_token
        - expires_in
        - token_type
        - user
...
    UserProfile:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
          example: 'user@example.com'
        first_name:
          type: string
          example: 'John'
        last_name:
          type: string
          example: 'Doe'
        role:
          type: string
          enum: [user, admin, manager, technician]
          example: 'user'
        email_verified:
          type: boolean
          example: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
      required:
        - id
        - email
        - role
        - email_verified
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: 'user@example.com'
      required:
        - email
    PasswordResetConfirm:
      type: object
      properties:
        token:
          type: string
          description: Password reset token from email
          example: 'abc123def456'
        new_password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: 'NewSecureP@ss456'
      required:
        - token
        - new_password
...
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: The refresh token obtained during login
          example: 'dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...'
      required:
        - refresh_token

    # --------------------------------------------------------------------------
    # Customer Schemas
    # --------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error:
              code: VALIDATION_ERROR
              message: 'Request validation failed'
              details:
                - field: email
                  message: 'Invalid email format'
                  code: INVALID_FORMAT
...
    RateLimited:
      description: Too many requests
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
          example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: 'Too many requests. Please try again later.'
              details:
                - retry_after: 60

# ==============================================================================
# Paths
# ==============================================================================
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: |
        Authenticates a user with email and password, returning JWT tokens.

        **Rate Limiting:** 5 requests per minute per IP address.

        **Account Locking:** After 5 failed attempts, the account is locked for 15 minutes.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: 'user@example.com'
              password: 'SecureP@ss123'
      responses:
        '200':
          description: Login successful
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
                refresh_token: 'dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...'
                expires_in: 3600
                token_type: 'Bearer'
                user:
                  id: '550e8400-e29b-41d4-a716-446655440000'
                  email: 'user@example.com'
                  first_name: 'John'
                  last_name: 'Doe'
                  role: 'user'
                  email_verified: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: 'Invalid email or password'
        '403':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: ACCOUNT_LOCKED
                  message: 'Account locked due to too many failed login attempts'
                  details:
                    - unlock_at: '2024-01-22T11:00:00Z'
                      minutes_remaining: 14
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access token.

        The refresh token is single-use and a new one will be provided with each refresh.
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: INVALID_REFRESH_TOKEN
                  message: 'Refresh token is invalid or expired'
...
  /auth/reset-request:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: |
        Sends a password reset email to the specified address if an account exists.

        **Rate Limiting:** 3 requests per hour per email address.

        For security, this endpoint always returns success even if the email is not registered.
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: 'If an account exists with this email, a password reset link has been sent.'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: |
        Resets the password using a token received via email.

        The reset token is valid for 1 hour.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: 'Password has been reset successfully. You can now login with your new password.'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: INVALID_RESET_TOKEN
                  message: 'Password reset token is invalid or expired'
...